+++ fiddles/vue/fiddle-0009-SkeletonLoaderJs/src/composables/useLocalStorage.js
@@ -1,0 +1,9 @@
+
+export default function useLocalStorage(key, defaultValue) {
+  const data = localStorage.getItem(key) && localStorage.getItem(key) !== 'null' ? JSON.parse(localStorage.getItem(key)) : defaultValue;
+  if (data) {
+    localStorage.setItem(key, JSON.stringify(data));
+  }
+  return data
+
+}